requirements.txt
asgiref==3.9.1
crispy-bootstrap5==2025.6
dj-database-url==3.0.1
Django==5.2.5
django-countries==7.6.1
django-crispy-forms==2.4
django-jazzmin==3.0.1
django-tinymce==4.1.0
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.9.0
